refactor : service 로직 수정 및 네이밍 수정
