+++ src/main/java/site/coach_coach/coach_coach_server/sport/dto/CoachingSportDto.java
@@ -1,15 +1,13 @@
 package site.coach_coach.coach_coach_server.sport.dto;
 
-
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
 @Builder
 public record CoachingSportDto(
-
-	@NotBlank
+	@NotNull
 	Long sportId,
-
 	@NotBlank
 	String sportName
 ) {
