+++ src/main/java/site/coach_coach/coach_coach_server/sport/dto/InterestedSportDto.java
@@ -1,13 +1,13 @@
 package site.coach_coach.coach_coach_server.sport.dto;
 
 import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.NotNull;
 import lombok.Builder;
 
 @Builder
 public record InterestedSportDto(
-	@NotBlank
+	@NotNull
 	Long userId,
-
 	@NotBlank
 	String sportName
 ) {
